move utils to dedicated file
