+++ pallets/proof-of-stake/src/tests.rs
@@ -7,6 +7,7 @@ use crate::mock::*;
 use frame_support::{assert_err, assert_err_ignore_postinfo, assert_ok};
 use mockall::predicate::eq;
 use serial_test::serial;
+use crate::utils::*;
 
 use mangata_support::traits::{ComputeIssuance, GetIssuance};
